refactor: move useBadges to features/badges
